pyproject.toml
[tool.poetry]
name = "chpqa"
version = "0.1.0"
description = ""
authors = ["Rhys Jones <rhys.jones@e2s.energy>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^1.5.3"
matplotlib = "^3.7.0"
ipykernel = "^6.21.2"
pylint = "^2.17.4"
streamlit = "^1.31.0"
fastapi = "^0.109.2"
pydantic-settings = "^2.1.0"
pyjwt = "^2.8.0"
uvicorn = "^0.27.1"
cryptography = "41.0.5"
authlib = "^1.3.0"
auth0-python = "^4.7.0"
mkdocs = "^1.5.3"
mkdocs-autorefs = "^1.0.1"
mkdocs-material = "^9.5.13"
mkdocs-material-extensions = "^1.3.1"
mkdocstrings = "^0.24.1"
mkdocstrings-python = "^1.9.0"

[tool.poetry.group.dev.dependencies]
pylint = "^2.16.2"
yapf = "^0.32.0"
pytest = "^7.2.1"
ipykernel = "^6.21.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
